python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentseal"
version = "0.6.1"
description = "Security toolkit for AI agents - machine scan for dangerous skills/MCP configs + prompt injection/extraction testing"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "FSL-1.1-Apache-2.0"}
authors = [
    {name = "AgentSeal", email = "hello@agentseal.org"},
]
keywords = ["security", "ai-agents", "llm", "prompt-injection", "pentesting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "httpx>=0.28",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
semantic = ["onnxruntime>=1.17", "tokenizers>=0.15", "numpy>=1.24", "huggingface-hub>=0.20"]
shield = ["watchdog>=4.0"]
all = ["openai>=1.0", "anthropic>=0.30", "pyyaml>=6.0", "onnxruntime>=1.17", "tokenizers>=0.15", "numpy>=1.24", "huggingface-hub>=0.20", "watchdog>=4.0"]

[project.urls]
Homepage = "https://agentseal.org"
Repository = "https://github.com/AgentSeal/agentseal"
Issues = "https://github.com/AgentSeal/agentseal/issues"

[tool.hatch.build.targets.wheel]
exclude = [
    "agentseal/probes/mcp_tools.py",
    "agentseal/probes/rag_poisoning.py",
    "agentseal/probes/multimodal.py",
    "agentseal/genome.py",
    "agentseal/report.py",
]

[project.scripts]
agentseal = "agentseal.cli:main"
